+++ src/main/java/com/lol/community/user/domain/User.java
@@ -13,7 +13,7 @@ public class User extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, length = 100)
     private String name;
@@ -28,7 +28,7 @@ public class User extends BaseEntity {
     private String grade;
 
     @Builder
-    public User (Integer id) {
+    public User (Long id) {
         this.id = id;
     }
 }
